+++ src/proxy_control.h
@@ -4,6 +4,7 @@
 #include "proxy.h"
 
 int proxy_control_init(proxy_t *p);
-int proxy_control_loop(proxy_t *p, int timeout_secs, int silent_secs);
+int proxy_control_loop(proxy_t *p, int timeout_secs, int silent_secs,
+                       int silent_exit_secs);
 
 #endif /* PROXY_CONTROL_H */